fix: remove laser weapon (rip my beloved) add flip animation like paper mario to character

+++ include/scenes/OptionsScene.h
@@ -16,7 +16,7 @@ public:
 private:
     TTF_Font* font = nullptr;
     int width, height;
-    int selectedIndex = 0;
+    size_t selectedIndex = 0;
 
     std::vector<std::string> options = {
         "Volume da Música",
